Implement chart pack algorithm
Chart pack algorithm: We guess a initial texture size based on the sum of chart areas, and try different methods then pick a best one, if pack failed because of texture size, we grow the size until it fit.

+++ simpleuv/meshdatatype.h
@@ -12,7 +12,7 @@ struct Vertex
 
 struct Face
 {
-    int indicies[3];
+    size_t indicies[3];
 };
 
 struct TextureCoord
